Split analysis/samples.rs into focused sub-modules (#597)
Split the monolithic samples.rs (1323 lines) into a directory with
focused sub-modules:

- samples/mod.rs — Public API, re-exports, core HelpfulSample type
- samples/gpu_types.rs — GPU-compatible #[repr(C)] data formats
- samples/statistics.rs — NeuronStats, HelpfulStats, HarmfulStats, ReluStats
- samples/thresholds.rs — Threshold computation and source variance analysis

Public API unchanged; all items re-exported from samples/mod.rs.
